Separate reporting boundaries from core authority
Move descriptive evidence/status implementations into reporting and ops homes while keeping core shims so existing CLI, runtime, and package import surfaces remain stable. The physical layout now matches the architecture boundary without widening runtime authority.

Constraint: Preserve public CLI/runtime behavior, package exports, React Web bounded path, and no RN/TUI/WebView support widening.

Rejected: Documentation-only cleanup | issue #676 requires physical separation after PR #675.

Rejected: Removing src/core deep paths | existing consumers and tests require compatibility shims.

Confidence: high

Scope-risk: moderate

Directive: Keep reporting/evidence/status surfaces descriptive; do not treat src/reporting or src/ops moves as runtime authority.

Tested: npm run build; targeted node --test for React Web/worktree/artifact/operator/architecture/physical-boundaries surfaces; npm test (690 passing).

Not-tested: External downstream packages that deep-import unpublished dist paths beyond the preserved core shims.
